dbeS cleanup

+++ dbeS/Bilder_xml.php
@@ -192,7 +192,7 @@ if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
  * @param array  $xml
  * @param string $unzipPath
  */
-function bearbeite($xml, $unzipPath)
+function bearbeite($xml, string $unzipPath)
 {
     $img_arr                 = mapArray($xml['bilder'], 'tartikelpict', $GLOBALS['mArtikelPict']);
     $kategoriebild_arr       = mapArray($xml['bilder'], 'tkategoriepict', $GLOBALS['mKategoriePict']);
@@ -370,7 +370,8 @@ function bearbeite($xml, $unzipPath)
             $Hersteller = Shop::Container()->getDB()->query(
                 "SELECT cSeo
                     FROM thersteller
-                    WHERE kHersteller = " . (int)$Herstellerbild->kHersteller, 1
+                    WHERE kHersteller = " . (int)$Herstellerbild->kHersteller,
+                \DB\ReturnType::SINGLE_OBJECT
             );
             if (isset($Hersteller->cSeo) && strlen($Hersteller->cSeo) > 0) {
                 $Herstellerbild->cPfad = str_replace("/", "_", $Hersteller->cSeo . '.' . $Bildformat);
@@ -637,9 +638,10 @@ function gibEigenschaftwertbildname($Eigenschaftwertbild, $Bildformat)
             : $Eigenschaftwertbild->cPfad . '.' . $Bildformat;
     }
     $Eigenschaftwert = Shop::Container()->getDB()->query(
-        "SELECT kEigenschaftWert, cArtNr, cName, kEigenschaft
+        'SELECT kEigenschaftWert, cArtNr, cName, kEigenschaft
             FROM teigenschaftwert
-            WHERE kEigenschaftWert = " . (int)$Eigenschaftwertbild->kEigenschaftWert, 1
+            WHERE kEigenschaftWert = ' . (int)$Eigenschaftwertbild->kEigenschaftWert,
+        \DB\ReturnType::SINGLE_OBJECT
     );
     $Bildname        = $Eigenschaftwert->kEigenschaftWert;
     if ($Eigenschaftwert->cName) {
@@ -658,9 +660,10 @@ function gibEigenschaftwertbildname($Eigenschaftwertbild, $Bildformat)
                             ON tseo.cKey = 'kArtikel'
                             AND tseo.kKey = tartikel.kArtikel
                             " . $cSQL . "
-                        WHERE teigenschaftwert.kEigenschaft=teigenschaft.kEigenschaft
+                        WHERE teigenschaftwert.kEigenschaft = teigenschaft.kEigenschaft
                             AND teigenschaft.kArtikel = tartikel.kArtikel
-                            AND teigenschaftwert.kEigenschaftWert = " . (int)$Eigenschaftwertbild->kEigenschaftWert, 1
+                            AND teigenschaftwert.kEigenschaftWert = " . (int)$Eigenschaftwertbild->kEigenschaftWert,
+                    \DB\ReturnType::SINGLE_OBJECT
                 );
                 if (!empty($Artikel->cArtNr) && !empty($Eigenschaftwert->cArtNr)) {
                     $Bildname = gibAusgeschriebeneUmlaute($Artikel->cArtNr) . '_' . gibAusgeschriebeneUmlaute($Eigenschaftwert->cArtNr);
@@ -677,12 +680,18 @@ function gibEigenschaftwertbildname($Eigenschaftwertbild, $Bildformat)
                             " . $cSQL . "
                         WHERE teigenschaftwert.kEigenschaft = teigenschaft.kEigenschaft
                             AND teigenschaft.kArtikel = tartikel.kArtikel
-                            AND teigenschaftwert.kEigenschaftWert = " . $Eigenschaftwertbild->kEigenschaftWert, 1
+                            AND teigenschaftwert.kEigenschaftWert = " . $Eigenschaftwertbild->kEigenschaftWert,
+                    \DB\ReturnType::SINGLE_OBJECT
                 );
 
-                $Eigenschaft = Shop::Container()->getDB()->query("SELECT cName FROM teigenschaft WHERE kEigenschaft = " . $Eigenschaftwert->kEigenschaft,
-                    1);
-                if ((!empty($Artikel->cSeo) || !empty($Artikel->cName)) && !empty($Eigenschaft->cName) && !empty($Eigenschaftwert->cName)) {
+                $Eigenschaft = Shop::Container()->getDB()->query(
+                    'SELECT cName FROM teigenschaft WHERE kEigenschaft = ' . $Eigenschaftwert->kEigenschaft,
+                    \DB\ReturnType::SINGLE_OBJECT
+                );
+                if ((!empty($Artikel->cSeo) || !empty($Artikel->cName))
+                    && !empty($Eigenschaft->cName)
+                    && !empty($Eigenschaftwert->cName)
+                ) {
                     if ($Artikel->cSeo) {
                         $Bildname = $Artikel->cSeo . '_' .
                             gibAusgeschriebeneUmlaute($Eigenschaft->cName) . '_' .
@@ -733,7 +742,8 @@ function gibKategoriebildname($Kategoriebild, $Bildformat)
                 ON tseo.cKey = 'kKategorie'
                 AND tseo.kKey = tkategorie.kKategorie
                 " . $cSQL . "
-            WHERE tkategorie.kKategorie = " . (int)$Kategoriebild->kKategorie, 1
+            WHERE tkategorie.kKategorie = " . (int)$Kategoriebild->kKategorie,
+        \DB\ReturnType::SINGLE_OBJECT
     );
     $Bildname  = $Kategoriebild->cPfad;
     if ($Kategorie->cName) {
@@ -797,7 +807,8 @@ function gibArtikelbildname($img, $Bildformat)
                 ON tseo.cKey = 'kArtikel'
                 AND tseo.kKey = tartikel.kArtikel
                 " . $cSQL . "
-            WHERE tartikel.kArtikel = " . (int)$img->kArtikel, 1
+            WHERE tartikel.kArtikel = " . (int)$img->kArtikel,
+        \DB\ReturnType::SINGLE_OBJECT
     );
     $Bildname = $img->cPfad;
     if ($Artikel->cName) {
@@ -885,7 +896,7 @@ function streicheSonderzeichen($str)
  * @param string $container
  * @return int
  */
-function erstelleThumbnailBranded($imgFilename, $zielbild, $breite, $hoehe, $quality = 80, $container = 'N')
+function erstelleThumbnailBranded($imgFilename, $zielbild, $breite, $hoehe, int $quality = 80, $container = 'N')
 {
     $vergroessern = 0;
     if ($GLOBALS['Einstellungen']['bilder']['bilder_skalieren'] === 'Y') {
@@ -1110,7 +1121,12 @@ function bearbeiteDeletes($xml)
         if (is_array($xml['del_bilder']['kHersteller'])) {
             foreach ($xml['del_bilder']['kHersteller'] as $kHersteller) {
                 if ((int)$kHersteller > 0) {
-                    Shop::Container()->getDB()->update('thersteller', 'kHersteller', (int)$kHersteller, (object)['cBildpfad' => '']);
+                    Shop::Container()->getDB()->update(
+                        'thersteller',
+                        'kHersteller',
+                        (int)$kHersteller,
+                        (object)['cBildpfad' => '']
+                    );
                     foreach (Shop::Container()->getDB()->selectAll('tartikel', 'kHersteller', (int)$kHersteller,
                         'kArtikel') as $article) {
                         $cacheTags[] = $article->kArtikel;
@@ -1118,8 +1134,12 @@ function bearbeiteDeletes($xml)
                 }
             }
         } elseif ((int)$xml['del_bilder']['kHersteller'] > 0) {
-            Shop::Container()->getDB()->update('thersteller', 'kHersteller', (int)$xml['del_bilder']['kHersteller'],
-                (object)['cBildpfad' => '']);
+            Shop::Container()->getDB()->update(
+                'thersteller',
+                'kHersteller',
+                (int)$xml['del_bilder']['kHersteller'],
+                (object)['cBildpfad' => '']
+            );
             foreach (Shop::Container()->getDB()->selectAll('tartikel', 'kHersteller', (int)$xml['del_bilder']['kHersteller'],
                 'kArtikel') as $article) {
                 $cacheTags[] = $article->kArtikel;
@@ -1166,13 +1186,12 @@ function bearbeiteDeletes($xml)
  * @param int      $kArtikelPict
  * @param int|null $nNr
  */
-function loescheArtikelPict($kArtikelPict, $nNr = null)
+function loescheArtikelPict(int $kArtikelPict, int $nNr = null)
 {
-    $kArtikelPict = (int)$kArtikelPict;
     if ($kArtikelPict > 0) {
         $oArtikelPict = null;
-        if ((int)$nNr > 0) {
-            $oArtikelPict = Shop::Container()->getDB()->select('tartikelpict', 'kArtikel', $kArtikelPict, 'nNr', (int)$nNr);
+        if ($nNr !== null && $nNr > 0) {
+            $oArtikelPict = Shop::Container()->getDB()->select('tartikelpict', 'kArtikel', $kArtikelPict, 'nNr', $nNr);
             $kArtikelPict = $oArtikelPict->kArtikelPict ?? 0;
         }
         deleteArticleImage(null, 0, $kArtikelPict);
@@ -1183,40 +1202,32 @@ function loescheArtikelPict($kArtikelPict, $nNr = null)
 }
 
 /**
- * @param int      $kKategoriePict
+ * @param int|null $kKategoriePict
  * @param int|null $kKategorie
  */
-function loescheKategoriePict($kKategoriePict, $kKategorie = null)
+function loescheKategoriePict(int $kKategoriePict = null, int $kKategorie = null)
 {
-    if ($kKategoriePict !== null) {
-        $kKategoriePict = (int)$kKategoriePict;
-        if ($kKategoriePict > 0) {
-            Shop::Container()->getDB()->delete('tkategoriepict', 'kKategoriePict', $kKategoriePict);
-            if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
-                Jtllog::writeLog('kKategoriePict geloescht: ' . $kKategoriePict, JTLLOG_LEVEL_DEBUG, false,
-                    'Bilder_xml');
-            }
+    if ($kKategoriePict !== null && $kKategoriePict > 0) {
+        Shop::Container()->getDB()->delete('tkategoriepict', 'kKategoriePict', $kKategoriePict);
+        if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
+            Jtllog::writeLog('kKategoriePict geloescht: ' . $kKategoriePict, JTLLOG_LEVEL_DEBUG, false,
+                'Bilder_xml');
         }
-    } else {
-        $kKategorie = (int)$kKategorie;
-        if ($kKategorie > 0) {
-            Shop::Container()->getDB()->delete('tkategoriepict', 'kKategorie', $kKategorie);
-            if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
-                Jtllog::writeLog('kKategoriePict geloescht: ' . $kKategorie, JTLLOG_LEVEL_DEBUG, false, 'Bilder_xml');
-            }
+    } elseif ($kKategorie !== null && $kKategorie > 0) {
+        Shop::Container()->getDB()->delete('tkategoriepict', 'kKategorie', $kKategorie);
+        if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
+            Jtllog::writeLog('kKategoriePict geloescht: ' . $kKategorie, JTLLOG_LEVEL_DEBUG, false, 'Bilder_xml');
         }
     }
 }
 
 /**
- * @param int $kEigenschaftwertPict
- * @param int $kEigenschaftwert
+ * @param int|null $kEigenschaftwertPict
+ * @param int|null $kEigenschaftwert
  */
-function loescheEigenschaftwertPict($kEigenschaftwertPict, $kEigenschaftwert = null)
+function loescheEigenschaftwertPict(int $kEigenschaftwertPict = null, int $kEigenschaftwert = null)
 {
-    $kEigenschaftwert     = (int)$kEigenschaftwert;
-    $kEigenschaftwertPict = (int)$kEigenschaftwertPict;
-    if ($kEigenschaftwert > 0) {
+    if ($kEigenschaftwert !== null && $kEigenschaftwert > 0) {
         Shop::Container()->getDB()->delete('teigenschaftwertpict', 'kEigenschaftWert', $kEigenschaftwert);
         if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
             Jtllog::writeLog('kEigenschaftWert geloescht: ' . $kEigenschaftwert, JTLLOG_LEVEL_DEBUG, false,
@@ -1224,7 +1235,7 @@ function loescheEigenschaftwertPict($kEigenschaftwertPict, $kEigenschaftwert = n
         }
     }
 
-    if ($kEigenschaftwertPict > 0) {
+    if ($kEigenschaftwertPict !== null && $kEigenschaftwertPict > 0) {
         Shop::Container()->getDB()->delete('teigenschaftwertpict', 'kEigenschaftwertPict', $kEigenschaftwertPict);
         if (Jtllog::doLog(JTLLOG_LEVEL_DEBUG)) {
             Jtllog::writeLog('kEigenschaftwertPict geloescht: ' . $kEigenschaftwertPict, JTLLOG_LEVEL_DEBUG, false,
@@ -1437,35 +1448,38 @@ function imagecopymerge_alpha_fast($dst_im, $src_im, $dst_x, $dst_y, $src_x, $sr
  * @param string $imgFilename
  * @return bool|string
  */
-function gibBildformat($imgFilename)
+function gibBildformat(string $imgFilename)
 {
-    if (file_exists($imgFilename)) {
-        $size = getimagesize($imgFilename);
-        $type = $size[2];
-        switch ($type) {
-            case IMAGETYPE_JPEG:
-                return 'jpg';
-                break;
+    if (!file_exists($imgFilename)) {
+        return false;
+    }
+    $size = getimagesize($imgFilename);
+    $type = $size[2];
+    switch ($type) {
+        case IMAGETYPE_JPEG:
+            return 'jpg';
+            break;
 
-            case IMAGETYPE_PNG:
-                if (function_exists('imagecreatefrompng')) {
-                    return 'png';
-                }
-                break;
+        case IMAGETYPE_PNG:
+            if (function_exists('imagecreatefrompng')) {
+                return 'png';
+            }
+            break;
 
-            case IMAGETYPE_GIF:
-                if (function_exists('imagecreatefromgif')) {
-                    return 'gif';
-                }
-                break;
+        case IMAGETYPE_GIF:
+            if (function_exists('imagecreatefromgif')) {
+                return 'gif';
+            }
+            break;
 
-            case IMAGETYPE_BMP:
-                if (function_exists('imagecreatefromwbmp')) {
-                    return 'bmp';
-                }
-                break;
-
-        }
+        case IMAGETYPE_BMP:
+            if (function_exists('imagecreatefromwbmp')) {
+                return 'bmp';
+            }
+            break;
+            
+        default:
+            break;
     }
 
     return false;
@@ -1479,29 +1493,26 @@ function gibBildformat($imgFilename)
  * @param int    $nContainerHeight
  * @return resource
  */
-function imageload_container($img, $nWidth, $nHeight, $nContainerWidth, $nContainerHeight)
+function imageload_container($img, int $nWidth, int $nHeight, $nContainerWidth, $nContainerHeight)
 {
     $imgInfo = getimagesize($img);
     switch ($imgInfo[2]) {
         case 1:
             $im = imagecreatefromgif($img);
             break;
-
         case 2:
             $im = imagecreatefromjpeg($img);
             break;
-
         case 3:
             $im = imagecreatefrompng($img);
             break;
-
         default:
             $im = imagecreatefromjpeg($img);
             break;
 
     }
 
-    if ($nWidth == 0 && $nHeight == 0) {
+    if ($nWidth === 0 && $nHeight === 0) {
         $nHeight = $imgInfo[1];
         $nWidth  = $imgInfo[0];
     }
@@ -1536,7 +1547,7 @@ function imageload_container($img, $nWidth, $nHeight, $nContainerWidth, $nContai
  * @param bool   $branding
  * @return resource
  */
-function imageload_alpha($img, $nWidth = 0, $nHeight = 0, $branding = false)
+function imageload_alpha($img, int $nWidth = 0, int $nHeight = 0, bool $branding = false)
 {
     $imgInfo = getimagesize($img);
     switch ($imgInfo[2]) {
@@ -1554,7 +1565,7 @@ function imageload_alpha($img, $nWidth = 0, $nHeight = 0, $branding = false)
 
     }
 
-    if ($nWidth == 0 && $nHeight == 0) {
+    if ($nWidth === 0 && $nHeight === 0) {
         $nHeight = $imgInfo[1];
         $nWidth  = $imgInfo[0];
     }
@@ -1597,7 +1608,7 @@ function imageload_alpha($img, $nWidth = 0, $nHeight = 0, $branding = false)
  * @param string $pfad
  * @return string
  */
-function neuerDateiname($pfad)
+function neuerDateiname(string $pfad)
 {
     $format = strtolower($GLOBALS['Einstellungen']['bilder']['bilder_dateiformat']);
     $pfad   = substr($pfad, 0, -3);
@@ -1609,51 +1620,27 @@ function neuerDateiname($pfad)
 /**
  * @param resource $im
  * @param string   $format
- * @param string   $pfad
+ * @param string   $path
  * @param int      $quality
  * @return bool
  */
-function speichereBild($im, $format, $pfad, $quality = 80)
+function speichereBild($im, $format, $path, int $quality = 80)
 {
     if (!$format || !$im) {
         return false;
     }
 
-    $pfad = neuerDateiname($pfad);
+    $path = neuerDateiname($path);
 
     switch (strtolower($format)) {
         case 'jpg':
-            if (!function_exists('imagejpeg')) {
-                return false;
-            }
-
-            return imagejpeg($im, $pfad, $quality);
-            break;
-
+            return function_exists('imagejpeg') ? imagejpeg($im, $path, $quality) : false;
         case 'png':
-            if (!function_exists('imagepng')) {
-                return false;
-            }
-
-            return imagepng($im, $pfad);
-            break;
-
+            return function_exists('imagepng') ? imagepng($im, $path) : false;
         case 'gif':
-            if (!function_exists('imagegif')) {
-                return false;
-            }
-
-            return imagegif($im, $pfad);
-            break;
-
+            return function_exists('imagegif') ? imagegif($im, $path) : false;
         case 'bmp':
-            if (!function_exists('imagewbmp')) {
-                return false;
-            }
-
-            return imagewbmp($im, $pfad);
-            break;
-
+            return function_exists('imagewbmp') ? imagewbmp($im, $path) : false;
     }
 
     return false;
@@ -1662,22 +1649,22 @@ function speichereBild($im, $format, $pfad, $quality = 80)
 /**
  * @return array
  */
-function holeBilderEinstellungen()
+function holeBilderEinstellungen(): array
 {
     $oBranding_arr    = [];
-    $oBrandingTMP_arr = Shop::Container()->getDB()->query("SELECT * FROM tbranding", 2);
-
-    if (is_array($oBrandingTMP_arr) && count($oBrandingTMP_arr) > 0) {
-        foreach ($oBrandingTMP_arr as $oBrandingTMP) {
-            $oBranding_arr[$oBrandingTMP->cBildKategorie] = $oBrandingTMP;
-        }
-        foreach ($oBranding_arr as $i => $oBranding) {
-            $oBranding_arr[$i]->oBrandingEinstellung = Shop::Container()->getDB()->select(
-                'tbrandingeinstellung',
-                'kBranding',
-                (int)$oBranding->kBranding
-            );
-        }
+    $oBrandingTMP_arr = Shop::Container()->getDB()->query(
+        'SELECT * FROM tbranding',
+        \DB\ReturnType::ARRAY_OF_OBJECTS
+    );
+    foreach ($oBrandingTMP_arr as $oBrandingTMP) {
+        $oBranding_arr[$oBrandingTMP->cBildKategorie] = $oBrandingTMP;
+    }
+    foreach ($oBranding_arr as $i => $oBranding) {
+        $oBranding_arr[$i]->oBrandingEinstellung = Shop::Container()->getDB()->select(
+            'tbrandingeinstellung',
+            'kBranding',
+            (int)$oBranding->kBranding
+        );
     }
 
     return $oBranding_arr;
